+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 2020.2.2
 
+#### Added
+
+* Added generic `bool Manager.TryGet<T>(out T manager)`
+
 #### Fixed
 
 * **BEHAVIOR CHANGE/FIX** : Managers are now initialized during `BeforeSceneLoad` both in editor and runtime to prevent behavior discrepancies and rare `Resource.Load<>()` issues.